[UPDATE] 종목당 시나리오 최대 1개 / 조건 추가시 BUY, SELL 당 최대 1개 제한
