+++ back/build.gradle
@@ -47,7 +47,7 @@ dependencies {
 	// scrapping
 	implementation 'org.jsoup:jsoup:1.14.3'
 
-	// Spring WebSocket
+	// websocket
 	implementation 'org.springframework.boot:spring-boot-starter-websocket'
 }
