+++ back/src/main/java/com/thered/stocksignal/config/WebSocketConfig.java
@@ -16,7 +16,7 @@ public class WebSocketConfig implements WebSocketConfigurer {
 
     @Override
     public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-        registry.addHandler(webSocketHandler, "/ws")
+        registry.addHandler(webSocketHandler, "/stocksignal")
                 .setAllowedOrigins("*");
     }
 }
